+++ core/controller/onetable.php
@@ -184,3 +184,4 @@ class onetableDocLister extends DocLister{
         return $out;
     }
 }
+?>