+++ core/controller/site_content_tags.php
@@ -374,3 +374,4 @@ class site_content_tagsDocLister extends DocLister{
 		return $out;
 	}
 }
+?>