+++ HISTORY.md
@@ -2,6 +2,7 @@
 * New method for template on DocLister::_getChunk
 * New extender template for easy add new custom method for template
 * fix url for main page
+* fix EOF
 
 ## 1.0.7 (2013-03-12)
 * add method DocLister::checkDL() for control main parameter